Move Favorites endpoints to beers controller, add InvalidIdMessage
